events: Move extensible event RoomMessageEventContent convertors
