+++ crates/ruma-common/src/events/room/message.rs
@@ -8,27 +8,6 @@ use ruma_macros::EventContent;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 
-#[cfg(feature = "unstable-msc3246")]
-use crate::events::audio::{AudioContent, AudioEventContent};
-#[cfg(feature = "unstable-msc3551")]
-use crate::events::file::{FileContent, FileEventContent};
-#[cfg(feature = "unstable-msc3552")]
-use crate::events::image::{ImageContent, ImageEventContent, ThumbnailContent};
-#[cfg(feature = "unstable-msc3553")]
-use crate::events::video::{VideoContent, VideoEventContent};
-#[cfg(feature = "unstable-msc3245")]
-use crate::events::voice::{VoiceContent, VoiceEventContent};
-#[cfg(feature = "unstable-msc1767")]
-use crate::events::{
-    emote::EmoteEventContent,
-    message::{MessageContent, MessageEventContent},
-    notice::NoticeEventContent,
-};
-#[cfg(feature = "unstable-msc3488")]
-use crate::{
-    events::location::{AssetContent, LocationContent, LocationEventContent},
-    MilliSecondsSinceUnixEpoch,
-};
 use crate::{
     serde::{JsonObject, StringEnum},
     OwnedEventId, PrivOwnedStr,
@@ -319,117 +298,6 @@ impl RoomMessageEventContent {
     /// Return a reference to the message body.
     pub fn body(&self) -> &str {
         self.msgtype.body()
-    }
-}
-
-#[cfg(feature = "unstable-msc3246")]
-impl From<AudioEventContent> for RoomMessageEventContent {
-    fn from(content: AudioEventContent) -> Self {
-        let AudioEventContent { message, file, audio, relates_to } = content;
-
-        Self {
-            msgtype: MessageType::Audio(AudioMessageEventContent::from_extensible_content(
-                message, file, audio,
-            )),
-            relates_to,
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc1767")]
-impl From<EmoteEventContent> for RoomMessageEventContent {
-    fn from(content: EmoteEventContent) -> Self {
-        let EmoteEventContent { message, relates_to, .. } = content;
-
-        Self { msgtype: MessageType::Emote(message.into()), relates_to }
-    }
-}
-
-#[cfg(feature = "unstable-msc3551")]
-impl From<FileEventContent> for RoomMessageEventContent {
-    fn from(content: FileEventContent) -> Self {
-        let FileEventContent { message, file, relates_to } = content;
-
-        Self {
-            msgtype: MessageType::File(FileMessageEventContent::from_extensible_content(
-                message, file,
-            )),
-            relates_to,
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc3552")]
-impl From<ImageEventContent> for RoomMessageEventContent {
-    fn from(content: ImageEventContent) -> Self {
-        let ImageEventContent { message, file, image, thumbnail, caption, relates_to } = content;
-
-        Self {
-            msgtype: MessageType::Image(ImageMessageEventContent::from_extensible_content(
-                message, file, image, thumbnail, caption,
-            )),
-            relates_to,
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc3488")]
-impl From<LocationEventContent> for RoomMessageEventContent {
-    fn from(content: LocationEventContent) -> Self {
-        let LocationEventContent { message, location, asset, ts, relates_to } = content;
-
-        Self {
-            msgtype: MessageType::Location(LocationMessageEventContent::from_extensible_content(
-                message, location, asset, ts,
-            )),
-            relates_to,
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc1767")]
-impl From<MessageEventContent> for RoomMessageEventContent {
-    fn from(content: MessageEventContent) -> Self {
-        let MessageEventContent { message, relates_to, .. } = content;
-
-        Self { msgtype: MessageType::Text(message.into()), relates_to }
-    }
-}
-
-#[cfg(feature = "unstable-msc1767")]
-impl From<NoticeEventContent> for RoomMessageEventContent {
-    fn from(content: NoticeEventContent) -> Self {
-        let NoticeEventContent { message, relates_to, .. } = content;
-
-        Self { msgtype: MessageType::Notice(message.into()), relates_to }
-    }
-}
-
-#[cfg(feature = "unstable-msc3553")]
-impl From<VideoEventContent> for RoomMessageEventContent {
-    fn from(content: VideoEventContent) -> Self {
-        let VideoEventContent { message, file, video, thumbnail, caption, relates_to } = content;
-
-        Self {
-            msgtype: MessageType::Video(VideoMessageEventContent::from_extensible_content(
-                message, file, video, thumbnail, caption,
-            )),
-            relates_to,
-        }
-    }
-}
-
-#[cfg(feature = "unstable-msc3245")]
-impl From<VoiceEventContent> for RoomMessageEventContent {
-    fn from(content: VoiceEventContent) -> Self {
-        let VoiceEventContent { message, file, audio, voice, relates_to } = content;
-
-        Self {
-            msgtype: MessageType::Audio(AudioMessageEventContent::from_extensible_voice_content(
-                message, file, audio, voice,
-            )),
-            relates_to,
-        }
     }
 }
